refactor: move nested circuit traces from _10_p into _11_circuits
Relocate endoscaling circuit trace creation (step_rxs, endoscalar_rx,
points_rx) from compute_p to compute_internal_circuits so that all
circuit tracing lives in _11_circuits. The P/NativeP structure is
preserved for future nested P computation. compute_p now returns the
PointsWitness and Uendo intermediates for _11_circuits to consume.
